Fix eaten grimmkin flames when using PlayerDataTracker
When upgrading Grimmchild with 4 flames and PDT enabled,
the game would set your flames to 0 (as normal), then rando
would set them to 1 (giving back the extra flame), and then
PDT would set it to 0 again. This is because the SetPlayerIntHook
requires the hook itself to write the new value even if it's not
overriding it.

The new way of doing this is, hopefully, more robust. It also,
as a bonus, makes PDT always show the correct number of flames
at all times.