Fix logic for Failed Champion and GPZ
Crossroads_10 (the False Knight room) is not infected in room rando
to prevent the left transition from being blocked, so the infection pogo
with wings is not possible there.

The entrance to Grey Prince Zote only appears if the player has Monarch
Wings. This allows us to simplify the GPZ logic, as alternative ways of
rescuing Bretta without Wings are no longer relevant.