Initial implementation of left/right claw
(Also fixed a stupid bug that accidentally got in from the previous commit)

Placeholders for left and right claw - one only works when facing left, the other when facing right.

Eventually, there will be an option to have a left claw and a right claw (instead of claw), both in mantis village near where claw usually is. (Neither will be duped in that case, as most of the time either claw will be enough)