Logic
Various changes to logic involving RIGHTCLAW based on the seed I just playtested.

Note: Climbing Deepnest_03 with only one claw is quite cursed - there's a jump below the darknest entrance that needs LEFTCLAW and a jump above that needs RIGHTCLAW. Accordingly it's possible to get all the way to failed tramway with just RIGHTCLAW when coming from dark deepnest (but not from the hot springs), though something more is needed to get to the bench.