Logic changes
Sanctum logic note: LEFTCLAW gets everything for free, and WINGS gets everything except for the first jump in the rewards room (which needs ANYCLAW too). With only RIGHTCLAW,  SUPERDASH is neeeded to hit the lever to unlock the soul warrior fight, and also to escape the soul master arena. MILDSKIPS is because there's a pogo (like the spire pogo) to get to the soul warrior lever, and it's definitely harder than the spire pogo. In practice, everything should be in easy (with RIGHTCLAW + SUPERDASH) if leverskips were in logic.

Getting to the top-left of Deepnest_East_07 can be done with ANYCLAW.