Replace claw pickup rather than destroy/create
I think it's probably safer to replace the old claw rather than destroy it and create a new shiny. The item is replaced with a new shiny because I like having both pickups on the ground rather than on the shelf.

(Also removed a few superfluous parentheses in this commit)