Add a randomly duped claw piece
One piece chosen randomly is duped; in this version, that means that the other claw piece is unique. Ideally collecting two of one piece would give the other, but that would require more effort than it's worth, given that there is likely to be a rewrite of this part of the code in the future.