Acid skips
Crossroads -> Fog Canyon with wings instead of rightclaw: This acid skip is a bit harder than the usual claw cdash version, but substantially easier than most of the other acid skips in logic.

Hive grub: acid skip to the grub in logic with dash and wings (or, logic for the transition in room/area rando). Also, the grub is now in logic with just anyclaw and spicyskips, to match the transition logic.