Temp fix for the disappearing Right Claw bug
Old bug: Right Claw pickup wouldn't appear when the following two conditions hold:
- The previous scene was a "boss scene", e.g. Fungus3_40_boss (Fungus_3_40 with dream nail in inventory)
- Mantis Claw is in inventory

This is an issue with the scene change functionality (the same one that sometimes causes stag bells not to appear), but is particularly bad here because the room below is a boss scene, so the bug happens frequently in area rando.

In this version we just create two shinies and Object.Destroy the old claw pickup - this seems to work. That said an eventual fix to the stag bell pickup would render this unnecessary, though I suppose it's probably better to implement it in a symmetric way like this one...

This commit also turns off the dupe claw piece