Abyss_02 logic
Moving from left to right with just leftclaw is too cursed for easy logic. Any other movement, or MILDSKIPS (to pogo the husks), or - implicitly - SPIKETUNNELS (to go under the two difficult columns) is enough.

We treat the Upper Basin waypoint as the left side of the room (with the geo rocks).

This logic change only affects split claw mode - leftclaw or wings is required to get up from the bottom transition.