Prevent certain items from destroying themselves before we've had a chance to replace them
This is a niche bug (only occurs if we enter the room, from a boss scene, while we have the item we're trying to replace), but it's affected me on at least two separate occasions so here's a fix for it.

I believe these are the only items which can have this bug affect them - others (such as collector's map) only destroy themselves based on a scene data check, which obviously can't be a problem because there's no way for that check to return true.