Deactivate replacing L/R shade cloak with Shade Cloak
The advantage of replacing Left/Right shade cloak with Shade Cloak (big item popup, grimmkin reward etc) is that it correctly shows the item given, leading to less confusion. However, it also destroys the information of which dash direction has been given; in a coop seed, knowing that the shade cloak is a left rather than a right can change how the partner plays (e.g. if the partner already has an accessible left dash but not a hard-to-reach right dash). I think for a mode like this, I don't mind expecting the player to know (or figure out) that shadowdash\dash is not split.