Allow boss essence to count as a progression item
Previously, with boss essence randomized, boss essence pickups did not correctly update the essenceLocations in the progression manager. This meant that items at seer could only count for progression if they were reachable only with whispering root pickups, and any items at seer with randomized cost greater than (482 - essence tolerance) would be non-progression (or dupe major) items placed at the end of the ordered locations list.

The content of this commit is the change to ItemManager; the change to LogicManager is not important.