Make the Shade Cloak geo rock fix apply generically rather than mention locations by name
I believe that these two, as well as L/R shade cloak, are currently the only locations with this property. That said, having the code not mention locations by name is better for forward-compatibility.

For some reason, simply modifying the space field in the GetPosition action when placing the rock doesn't fix the issue :/