Remove Pogo-only ogres from easy/medium logic
The logic for Main_Fungal_Wastes not from Left_Fungal_Wastes technically covers only the right side of Fungus2_07; however, this can only be a problem in area rando with no movement, and even in that case it's possible to go round through Shrumal Ogres (if mildly annoying). For now, I don't think this is a high priority.