Add "Blitz" Mode for focusing progression into limited areas
During randomisation, some areas are selected, and all other areas are
flagged as "junk". Then before items are normally placed, all junk areas
are filled with non-progression items. Once this is complete, "normal"
placement occurs.

All items defined as being progression are placed in non-junk areas,
even if those items are not actually required in that seed (e.g. Tram
Pass when no progression items require it, or all grubs even if nothing is
at grubs).

Information about which areas are required can be found in the "Tracker"
in the inventory during play.